requirements.txt
openai
python-dotenv
pyyaml
ddgs
AsyncDDGS
playwright
beautifulsoup4
html2text